feat: ucan-key-support: add P-256 key support (aka ESRSA, aka secp256r1) (#46)
While it is possible for calling code to add arbitrary key support by
implementing the KeyMaterial trait, it is helpful to have some popular
key types included in the `ucan-key-support` crate.

The inclusion of P-256 is motivated by the support for that curve in the
'ts-ucan' npm library. Adding P-256 to `ucan-key-support` will mean the
same set of keys are supported by both libraries out of the box.

This commit adds support for this key type via the `p256` crate,
maintained by RustCrypto. These are the same maintainers as the 'rsa'
crate, and several recursive dependencies are shared.

The implementation is basically directly copied from the 'ed25519'
module.

Signed-off-by: bryan newbold <bnewbold@robocracy.org>