refactor: gate static plugin registration behind the target
There's likely a better way to do this, so that we're able to
disable the feature if desired, or for other environments, like
in no-std embedded targets. A feature could help with me, but I'd
like to have it enabled by default, and default-features brings
with it some other ergonomics concerns, so this is as good
a starting point as any.